Add a method to format a card as a 2 character string.
Notice I have to turn the suit from a symbol into as string to get at
the first character, and I can use the same indexing syntax on a
string as I can on an array.

I've tested a few more possibilities here; you don't need to be
exhaustive in your tsting, but this felt like it benefited from more
assertions.