Create a Hand class to represent a player's current hand.
My instinct to do this rather than just use an Array is because hands
will have logic associated with them - deciding whether they are bust
or not for example.

Notice that my test is just dealing with the simple case first - where
the rank of a card is the same as the number of points it is
worth. Next up will be dealing with picture cards.