Adding logic to deal with picture card scoring.
I'm not happy with this though - the scoring method is getting
complicated, and I'm violating a principle called "Tell, don't ask",
which means that I'm asking a card for some attribute about itself,
and then using an if statement to do different things based on the
result. In general, it is better to move this sort of logic onto the
object your asking about, and then just tell it to give you a
result (in this case the score).