Deal with Aces being 1 if necessary.
Add an ace? query method to card to see if a card is an ace or not. In
Hand#score we keep track of the number of aces we've seen in the hand,
as well as the total. Then, if the hand score is over the maximum, we
reduce the hand by 10 (11 - 1) until we're under the limit or we've
run out of aces we can do that for.

In the test for Card#ace? you'll see 2 new types of assertions: assert
and refute. These are the same as assert_equal true, whatever and
assert_equal false, whatever