Spike out a game and create a bin file to run it.
By spiking, I mean writing out some code but not testing it, and not
caring too much for the quality. The Game#run method is hideous at the
moment, because it is doing far too much. I'm writing it like this
because now I have Card and Hand, I'm not too sure how to proceed with
designing the game logic, so I just write out what I want to happen
procedurally, and rely on the fact I can refactor to improve the
design later.

I've created a new directory called bin/, and under this created a
small script that sets up the load path and starts the game off. The
first line of this file tells the console how to run this file, and I
make it executable by running the command line command:

  chmod +x bin/blackjack

See man chmod for more details. This allows me to just run

  bin/blackjack

from the command line, instead of

  ruby bin/blackjack