Extract out some of the UI logic from calculating the winner.
One of the big problems with the run method at the moment is that UI
logic (in this case communicating through the console using gets and
puts) is mixed up with the game logic. It is good practice in general
to keep these things separate, so I'm starting to tease them apart.